[DEF70678] Fix warning from RHEL 6.1 Beta. - Fix up warnings for 32 bit
[git-p4: depot-paths = "//dev/talpa/main/": change = 2107173]

+++ src/components/core/intercept_filters_impl/vetting_ctrl/vetting_ctrl.c
@@ -603,7 +603,7 @@ static void examineFile(const void* self, IEvaluationReport* report, const IPers
         return;
     }
 
-    dbg("[intercepted %u-%u-%u] allocated %lu bytes at 0x%p for vetting details", processParentPID(current), current->tgid, current->pid, sizeof(VettingDetails), details);
+    dbg("[intercepted %u-%u-%u] allocated %lu bytes at 0x%p for vetting details", processParentPID(current), current->tgid, current->pid, (long unsigned int) sizeof(VettingDetails), details);
 
     /* See how much memory do we need for VettingDetails packet */
     rootdir = threadInfo->rootDir(threadInfo);
@@ -881,7 +881,7 @@ static void examineFilesystem(const void* self, IEvaluationReport* report,
         return;
     }
 
-    dbg("[intercepted %u-%u-%u] allocated %lu bytes at 0x%p for vetting details", processParentPID(current), current->tgid, current->pid, sizeof(VettingDetails), details);
+    dbg("[intercepted %u-%u-%u] allocated %lu bytes at 0x%p for vetting details", processParentPID(current), current->tgid, current->pid, (long unsigned int) sizeof(VettingDetails), details);
 
     /* See how much memory do we need for VettingDetails packet */
     rootdir = threadInfo->rootDir(threadInfo);
